Make git-setup smarter by giving it some answers

+++ bin/ps1/git-setup.ps1
@@ -31,59 +31,72 @@ if (!$loc) {
 }
 Write-Host "Found git repository at $loc"
 
-#if (!(Test-Path "$loc\hooks\post-commit")) {
+
+while ((Read-Host "Insert post-commit hook? (Y/n)") -ne "n") {  # just one time, allows us to break out
     Write-Host "Creating git post-commit hook"
     $qtRepoToolsLoc = Get-LocalOrParentPath("qtrepotools")
     if (!$qtRepoToolsLoc) {
         Write-Host "Could not find qtrepotools\git-hooks\git_post_commit_hook"
-    } else {
-        # mingw-ize the path to the hook so that bash can find it
-        $drive = $qtRepoToolsLoc[0]
-        $hooksDir = "/$drive/" + $qtRepoToolsLoc.substring(3) + "/git-hooks"
-        $hooksDir = $hooksDir.replace("\", "/")
-        $hooksDir = $hooksDir.replace("\", "/")
-        $content =
+        $qtRepoToolsLoc = Read-Host "Please give location to qtrepotools"
+        if (!(Test-Path "$qtRepoToolsLoc\.git")) {
+            Write-Host "Could not install ($qtRepoToolsLoc is not a valid Git repo)"
+            break
+        }
+    }
+    
+    # mingw-ize the path to the hook so that bash can find it
+    $drive = $qtRepoToolsLoc[0]
+    $hooksDir = "/$drive/" + $qtRepoToolsLoc.substring(3) + "/git-hooks"
+    $hooksDir = $hooksDir.replace("\", "/")
+    $hooksDir = $hooksDir.replace("\", "/")
+    $content =
 @"
 #!/bin/sh
 export PATH=`$PATH:$hooksDir
 exec $hooksDir/git_post_commit_hook
 "@
         Set-Content "$loc\hooks\post-commit" $content
-    }
-#}
+        Write-Host "$loc\hooks\post-commit"
+    break
+}
 
-$copyCommitMsg = $false
-if ((Test-Path "$loc\hooks\commit-msg")) {
-    $commitMsgSize = (Get-ItemProperty "$loc\hooks\commit-msg").Length
-    if ($commitMsgSize -lt 200) {
+if ((Read-Host "Install commit-msg hook (for gerrit)? (Y/n)") -ne "n") {
+    $copyCommitMsg = $false
+    if ((Test-Path "$loc\hooks\commit-msg")) {
+        $commitMsgSize = (Get-ItemProperty "$loc\hooks\commit-msg").Length
+        if ($commitMsgSize -lt 200) {
+            $copyCommitMsg = $true
+        }
+    } else {
         $copyCommitMsg = $true
     }
-} else {
-    $copyCommitMsg = $true
+
+    if ($copyCommitMsg) {
+        Write-Host "Copying git commit-msg hook"
+        Push-Location "$loc\hooks\"
+        scp -p smd@codereview.qt-project.org:hooks/commit-msg .
+        Pop-Location
+    }
 }
 
-if ($copyCommitMsg) {
-    Write-Host "Copying git commit-msg hook"
-    Push-Location "$loc\hooks\"
-    scp -p smd@codereview.qt-project.org:hooks/commit-msg .
-    Pop-Location
+
+if ((Read-Host "Configure global .gitconfig? (y/N)") -eq "y") {
+    Write-Host "Configuring global .gitconfig"
+
+    if ("$env:COMPUTERNAME" -eq "PILSEN") {
+        GitConfigSet "url.qtsoftware:.insteadOf" "git@scm.dev.troll.no:"
+    }
+
+    GitConfigSet "mailmap.file" "t:\dev\devtools\aliases\mailmap"
+    GitConfigSet "alias.loginternal" "log --pretty=tformat:'commit %H%nAuthor: %an <%ae>%n (int): %aN <%aE>%nDate: %ad%n%n%s%n%n%b'"
+    GitConfigSet "core.autocrlf" "true"
+    GitConfigSet "alias.br" "branch"
+    GitConfigSet "alias.st" "status"
+    GitConfigSet "alias.co" "checkout"
+    GitConfigSet "alias.ce" "config --global -e"
+    GitConfigSet "alias.pushgerrit" "push gerrit HEAD:refs/for/master"
+
+    # Show resulting configuration
+    git config --global --list
+
 }
-
-Write-Host "Configuring global .gitconfig"
-
-if ("$env:COMPUTERNAME" -eq "PILSEN") {
-	GitConfigSet "url.qtsoftware:.insteadOf" "git@scm.dev.troll.no:"
-}
-
-GitConfigSet "mailmap.file" "t:\dev\devtools\aliases\mailmap"
-GitConfigSet "alias.loginternal" "log --pretty=tformat:'commit %H%nAuthor: %an <%ae>%n (int): %aN <%aE>%nDate: %ad%n%n%s%n%n%b'"
-GitConfigSet "core.autocrlf" "true"
-GitConfigSet "alias.br" "branch"
-GitConfigSet "alias.st" "status"
-GitConfigSet "alias.co" "checkout"
-GitConfigSet "alias.ce" "config --global -e"
-GitConfigSet "alias.pushgerrit" "push gerrit HEAD:refs/for/master"
-
-# Show resulting configuration
-git config --global --list
-
